Fix HDL WCI adress width when sizeOfConfigSpace is overridden to be 2^32/4GB core/local/util/ezxml/src/OcpiUtilEzxml.cxx: allow 'g' multiplier for 32 bit numbers. tools/cdk/ocpigen/ocpigen/include/wip.h: use 64 bits for sizeOfConfigSpace tools/cdk/ocpigen/ocpigen/src/wip_HDL.cxx: ditto, and tweak signal field widths for prettier "defs", and   suppress defining offsets for properties that are compile-time parameters tools/cdk/ocpigen/ocpigen/src/wip_parse.cxx: use 64 bits for sizeofConfigSpace, and    move sizeOfConfigSpace override AFTER all properties are parsed

+++ tools/cdk/ocpigen/ocpigen/include/wip.h
@@ -277,7 +277,7 @@ struct Property {
 #endif
 
 struct Control {
-  unsigned sizeOfConfigSpace;
+  uint64_t sizeOfConfigSpace;
   bool writableConfigProperties;
   bool readableConfigProperties;
   bool sub32BitConfigProperties;
